KU-104/refactor: 파티 관련 폴더 구조 정리
