+++ app/main-screens/main-p.tsx
@@ -5,7 +5,7 @@ import type { MainStackParamList } from "@/app/navigation/MainStack";
 import PartyHeader from "@/components/bars/PartyHeader";
 import PartySortFilterBar, { SortOption, FilterOption } from "@/components/bars/PartySortFilterBar";
 import PartyList from "@/components/lists/PartyList";
-import PartyBottomSheet from "@/components/lists/PartyBottomSheet";
+import PartyBottomSheet from "@/components/sections/PartyBottomSheet";
 
 import type { PartyItem } from "@/types/party";
 import { parties } from "@/mocks/parties";
